+++ crates/widgets/src/text_box.rs
@@ -1,10 +1,16 @@
-use std::cell::Cell;
-
 use super::behaviors::MouseBehavior;
 use crate::{
     prelude::*,
     shell::{Key, KeyEvent},
 };
+
+// --- KEYS --
+
+pub static ELEMENT_TEXT_BOX: &'static str = "text_box";
+
+static ID_CURSOR: &'static str = "id_cursor";
+
+// --- KEYS --
 
 #[derive(Clone)]
 enum TextBoxAction {
@@ -13,156 +19,267 @@ enum TextBoxAction {
 }
 
 /// The `TextBoxState` handles the text processing of the `TextBox` widget.
-#[derive(AsAny)]
+#[derive(Default, AsAny)]
 pub struct TextBoxState {
-    action: RefCell<Option<TextBoxAction>>,
-    cursor_x: Cell<f64>,
-    len: Cell<usize>,
-}
-
-impl Default for TextBoxState {
-    fn default() -> Self {
-        TextBoxState {
-            action: RefCell::new(None),
-            cursor_x: Cell::new(0.0),
-            len: Cell::new(0),
-        }
-    }
+    action: Option<TextBoxAction>,
+    len: usize,
+    cursor: Entity,
 }
 
 impl TextBoxState {
-    fn action(&self, action: TextBoxAction) {
-        *self.action.borrow_mut() = Some(action);
+    fn action(&mut self, action: TextBoxAction) {
+        self.action = Some(action);
     }
 
-    fn handle_key_event(&self, key_event: KeyEvent, ctx: &mut Context<'_>) {
+    fn handle_key_event(&mut self, key_event: KeyEvent, ctx: &mut Context<'_>) {
         if !ctx.widget().get::<bool>("focused") {
             return;
         }
 
-        let text = ctx.widget().clone::<String16>("text");
-        let mut current_selection = *ctx.child("cursor").get::<TextSelection>("text_selection");
-
         match key_event.key {
             Key::Left => {
-                if let Some(selection) = ctx
-                    .child("cursor")
-                    .try_get_mut::<TextSelection>("text_selection")
-                {
-                    selection.start_index =
-                        (current_selection.start_index as i32 - 1).max(0) as usize;
-                }
+                self.move_cursor_left(ctx);
             }
             Key::Right => {
-                if let Some(selection) = ctx
-                    .child("cursor")
-                    .try_get_mut::<TextSelection>("text_selection")
-                {
-                    selection.start_index = (current_selection.start_index + 1).min(text.len());
-                }
+                self.move_cursor_right(ctx);
             }
             Key::Backspace => {
-                if !text.is_empty() && current_selection.start_index > 0 {
-                    for _ in 0..=current_selection.length {
-                        ctx.widget()
-                            .get_mut::<String16>("text")
-                            .remove(current_selection.start_index - 1);
-                        current_selection.start_index =
-                            (current_selection.start_index as i32 - 1).max(0) as usize;
-                    }
-
-                    if let Some(selection) = ctx
-                        .child("cursor")
-                        .try_get_mut::<TextSelection>("text_selection")
-                    {
-                        selection.start_index = current_selection.start_index;
-                    }
-
-                    ctx.push_event_strategy_by_entity(
-                        ChangedEvent(ctx.entity),
-                        ctx.entity,
-                        EventStrategy::Direct,
-                    );
-                }
+                self.back_space(ctx);
             }
             Key::Delete => {
-                if !text.is_empty() && text.len() > current_selection.start_index {
-                    for _ in 0..=current_selection.length {
-                        ctx.widget()
-                            .get_mut::<String16>("text")
-                            .remove(current_selection.start_index);
-                    }
-
-                    ctx.push_event_strategy_by_entity(
-                        ChangedEvent(ctx.entity),
-                        ctx.entity,
-                        EventStrategy::Direct,
-                    );
-                }
+                self.delete(ctx);
             }
             Key::Enter => {
-                if *ctx.widget().get::<bool>("lost_focus_on_activation") {
-                    ctx.window().get_mut::<Global>("global").focused_widget = None;
-                    ctx.widget().set("focused", false);
-                    ctx.widget().update_theme_by_state(false);
+                self.activate(ctx);
+            }
+            Key::A(..) => {
+                // if cfg!(mac_os) {
+                //     if ctx
+                //         .window()
+                //         .get::<Global>("global")
+                //         .keyboard_state
+                //         .is_home_down()
+                //     {
+                //         self.select_all(ctx);
+                //     } else {
+                //         self.insert_char(key_event, ctx);
+                //     }
+                // } else {
+                if ctx
+                    .window()
+                    .get::<Global>("global")
+                    .keyboard_state
+                    .is_ctrl_down()
+                {
+                    self.select_all(ctx);
+                } else {
+                    self.insert_char(key_event, ctx);
                 }
-
-                ctx.push_event_strategy_by_entity(
-                    ActivateEvent(ctx.entity),
-                    ctx.entity,
-                    EventStrategy::Direct,
-                )
+                // }
             }
             _ => {
-                if key_event.text.is_empty() {
-                    return;
-                }
-                ctx.widget()
-                    .get_mut::<String16>("text")
-                    .insert_str(current_selection.start_index, key_event.text.as_str());
-
-                if let Some(selection) = ctx
-                    .child("cursor")
-                    .try_get_mut::<TextSelection>("text_selection")
-                {
-                    selection.start_index =
-                        current_selection.start_index + key_event.text.encode_utf16().count();
-                }
-
-                ctx.push_event_strategy_by_entity(
-                    ChangedEvent(ctx.entity),
-                    ctx.entity,
-                    EventStrategy::Direct,
-                );
+                self.insert_char(key_event, ctx);
             }
         }
     }
 
     fn request_focus(&self, ctx: &mut Context<'_>) {
-        let focused_widget = ctx.window().get::<Global>("global").focused_widget;
+        ctx.push_event_by_window(FocusEvent::RequestFocus(ctx.entity));
 
-        if (focused_widget.is_some() && focused_widget.unwrap() == ctx.entity)
-            || !ctx.widget().get::<bool>("enabled")
-        {
-            return;
+        if ctx.widget().get::<String16>("text").len() > 0 {
+            self.select_all(ctx);
         }
-
-        if let Some(old_focused_element) = ctx.window().get::<Global>("global").focused_widget {
-            let mut old_focused_element = ctx.get_widget(old_focused_element);
-            old_focused_element.set("focused", false);
-            old_focused_element.update_theme_by_state(false);
-        }
-
-        ctx.window().get_mut::<Global>("global").focused_widget = Some(ctx.entity);
-
-        ctx.widget().set("focused", true);
-        ctx.widget().update_theme_by_state(false);
     }
 
     // Reset selection and offset if text is changed from outside
     fn reset(&self, ctx: &mut Context<'_>) {
         ctx.widget().set("text_selection", TextSelection::default());
-        ctx.widget().set("scroll_offset", Point::default());
+        ctx.push_event_strategy_by_entity(
+            ChangedEvent(ctx.entity),
+            ctx.entity,
+            EventStrategy::Direct,
+        );
+    }
+
+    fn check_outside_update(&self, ctx: &mut Context<'_>) {
+        let len = ctx.widget().get::<String16>("text").len();
+        if self.len != len && self.len > len {
+            self.reset(ctx);
+        }
+    }
+
+    fn select_all(&self, ctx: &mut Context) {
+        let len = ctx.widget().get::<String16>("text").len();
+        ctx.widget()
+            .get_mut::<TextSelection>("text_selection")
+            .start_index = 0;
+        ctx.widget()
+            .get_mut::<TextSelection>("text_selection")
+            .length = len;
+    }
+
+    fn move_cursor_left(&mut self, ctx: &mut Context) {
+        if *ctx.get_widget(self.cursor).get::<bool>("expanded") {
+            if let Some(selection) = ctx
+                .get_widget(self.cursor)
+                .try_get_mut::<TextSelection>("text_selection")
+            {
+                selection.start_index = 0;
+                selection.length = 0;
+            }
+        }
+
+        if let Some(selection) = ctx
+            .get_widget(self.cursor)
+            .try_get_mut::<TextSelection>("text_selection")
+        {
+            selection.start_index = (selection.start_index as i32 - 1).max(0) as usize;
+            selection.length = 0;
+        }
+    }
+
+    fn move_cursor_right(&mut self, ctx: &mut Context) {
+        let text_len = ctx.widget().get::<String16>("text").len();
+
+        if *ctx.get_widget(self.cursor).get::<bool>("expanded") {
+            if let Some(selection) = ctx
+                .get_widget(self.cursor)
+                .try_get_mut::<TextSelection>("text_selection")
+            {
+                selection.start_index = text_len;
+                selection.length = 0;
+            }
+
+            return;
+        }
+
+        if let Some(selection) = ctx
+            .get_widget(self.cursor)
+            .try_get_mut::<TextSelection>("text_selection")
+        {
+            if selection.start_index < text_len {
+                selection.start_index = (selection.start_index + 1).min(text_len);
+            }
+            selection.length = 0;
+        }
+    }
+
+    fn clear_selection(&mut self, ctx: &mut Context) {
+        let selection = ctx.widget().clone::<TextSelection>("text_selection");
+
+        if let Some(text) = ctx.widget().try_get_mut::<String16>("text") {
+            for i in (selection.start_index..(selection.start_index + selection.length)).rev() {
+                text.remove(i);
+            }
+        }
+
+        ctx.widget()
+            .get_mut::<TextSelection>("text_selection")
+            .length = 0;
+    }
+
+    fn back_space(&mut self, ctx: &mut Context) {
+        let mut changed = false;
+
+        if *ctx.get_widget(self.cursor).get::<bool>("expanded") {
+            self.clear_selection(ctx);
+            changed = true;
+        } else {
+            let index = ctx
+                .widget()
+                .clone::<TextSelection>("text_selection")
+                .start_index;
+            if index > 0 {
+                ctx.widget().get_mut::<String16>("text").remove(index - 1);
+                ctx.widget()
+                    .get_mut::<TextSelection>("text_selection")
+                    .start_index = index - 1;
+
+                changed = true;
+            }
+        }
+
+        if changed {
+            ctx.push_event_strategy_by_entity(
+                ChangedEvent(ctx.entity),
+                ctx.entity,
+                EventStrategy::Direct,
+            );
+        }
+    }
+
+    fn delete(&mut self, ctx: &mut Context) {
+        let mut changed = false;
+
+        if *ctx.get_widget(self.cursor).get::<bool>("expanded") {
+            self.clear_selection(ctx);
+        } else {
+            let index = ctx
+                .widget()
+                .clone::<TextSelection>("text_selection")
+                .start_index;
+            if index < ctx.widget().get::<String16>("text").len() {
+                ctx.widget().get_mut::<String16>("text").remove(index);
+                changed = true;
+
+                ctx.widget()
+                    .get_mut::<TextSelection>("text_selection")
+                    .start_index = index;
+            }
+        }
+
+        if changed {
+            ctx.push_event_strategy_by_entity(
+                ChangedEvent(ctx.entity),
+                ctx.entity,
+                EventStrategy::Direct,
+            );
+        }
+    }
+
+    fn activate(&self, ctx: &mut Context) {
+        if *ctx.widget().get::<bool>("lost_focus_on_activation") {
+            ctx.push_event_by_window(FocusEvent::RemoveFocus(ctx.entity));
+        }
+
+        ctx.push_event_strategy_by_entity(
+            ActivateEvent(ctx.entity),
+            ctx.entity,
+            EventStrategy::Direct,
+        )
+    }
+
+    fn insert_char(&mut self, key_event: KeyEvent, ctx: &mut Context) {
+        if key_event.text.is_empty() {
+            return;
+        }
+
+        if *ctx.get_widget(self.cursor).get::<bool>("expanded") {
+            ctx.widget().set("text", String16::from(key_event.text));
+            if let Some(selection) = ctx
+                .get_widget(self.cursor)
+                .try_get_mut::<TextSelection>("text_selection")
+            {
+                selection.start_index = 1;
+                selection.length = 0
+            }
+        } else {
+            let current_selection = *ctx
+                .get_widget(self.cursor)
+                .get::<TextSelection>("text_selection");
+            ctx.widget()
+                .get_mut::<String16>("text")
+                .insert_str(current_selection.start_index, key_event.text.as_str());
+
+            if let Some(selection) = ctx
+                .get_widget(self.cursor)
+                .try_get_mut::<TextSelection>("text_selection")
+            {
+                selection.start_index =
+                    current_selection.start_index + key_event.text.encode_utf16().count();
+            }
+        }
+
         ctx.push_event_strategy_by_entity(
             ChangedEvent(ctx.entity),
             ctx.entity,
@@ -173,17 +290,16 @@ impl TextBoxState {
 
 impl State for TextBoxState {
     fn init(&mut self, _: &mut Registry, ctx: &mut Context<'_>) {
-        self.len.set(ctx.widget().get::<String16>("text").len());
+        self.cursor = ctx
+            .entity_of_child(ID_CURSOR)
+            .expect("TextBoxState.init: cursor child could not be found.");
+        self.len = ctx.widget().get::<String16>("text").len();
     }
 
     fn update(&mut self, _: &mut Registry, ctx: &mut Context<'_>) {
-        // check if text len is changed from outside
-        let len = ctx.widget().get::<String16>("text").len();
-        if self.len.get() != len && self.len.get() > len {
-            self.reset(ctx);
-        }
+        self.check_outside_update(ctx);
 
-        if let Some(action) = self.action.borrow().clone() {
+        if let Some(action) = self.action.clone() {
             match action {
                 TextBoxAction::Key(event) => {
                     self.handle_key_event(event, ctx);
@@ -194,62 +310,16 @@ impl State for TextBoxState {
             }
         }
 
-        *self.action.borrow_mut() = None;
+        self.action = None;
         ctx.widget().update_theme_by_state(false);
-        self.len.set(ctx.widget().get::<String16>("text").len());
-    }
-
-    fn update_post_layout(&mut self, _: &mut Registry, ctx: &mut Context<'_>) {
-        let mut cursor_x_delta = 0.0;
-        let mut scroll_viewer_width = 0.0;
-
-        {
-            if let Some(bounds) = ctx
-                .child("scroll_viewer")
-                .try_get_mut::<Rectangle>("bounds")
-            {
-                scroll_viewer_width = bounds.width();
-            }
-        }
-
-        // maybe not use scroll viewer here
-
-        // Adjust offset of text and cursor if cursor position is out of bounds
-
-        {
-            let mut cursor = ctx.child("cursor");
-
-            if let Some(margin) = cursor.try_get_mut::<Thickness>("margin") {
-                if margin.left() < 0.0 || margin.left() > scroll_viewer_width {
-                    cursor_x_delta = self.cursor_x.get() - margin.left();
-                    margin.set_left(self.cursor_x.get());
-                }
-                self.cursor_x.set(margin.left());
-            }
-
-            if let Some(bounds) = cursor.try_get_mut::<Rectangle>("bounds") {
-                bounds.set_x(self.cursor_x.get());
-            }
-        }
-
-        if cursor_x_delta != 0.0 {
-            {
-                if let Some(bounds) = ctx.child("text_block").try_get_mut::<Rectangle>("bounds") {
-                    bounds.set_x(bounds.x() + cursor_x_delta);
-                }
-            }
-
-            if let Some(scroll_offset) = ctx.widget().try_get_mut::<Point>("scroll_offset") {
-                scroll_offset.x += cursor_x_delta;
-            }
-        }
+        self.len = ctx.widget().get::<String16>("text").len();
     }
 }
 
 widget!(
     /// The `TextBox` widget represents a single line text input widget.
     ///
-    /// * CSS element: `text-box`
+    /// * CSS element: `text_box`
     TextBox<TextBoxState>: ActivateHandler, ChangedHandler, KeyDownHandler {
         /// Sets or shares the text property.
         text: String16,
@@ -284,12 +354,6 @@ widget!(
         /// Sets or shares the padding property.
         padding: Thickness,
 
-        /// Sets or shares the text offset property.
-        scroll_offset: Point,
-
-        /// Sets or shares the (wheel, scroll) delta property.
-        delta: Point,
-
         /// Sets or shares the focused property.
         focused: bool,
 
@@ -300,14 +364,23 @@ widget!(
 
 impl Template for TextBox {
     fn template(self, id: Entity, ctx: &mut BuildContext) -> Self {
+        let text_block = TextBlock::create()
+            .vertical_alignment("center")
+            .horizontal_alignment("start")
+            .foreground(id)
+            .text(id)
+            .water_mark(id)
+            .font(id)
+            .font_size(id)
+            .build(ctx);
+
         self.name("TextBox")
-            .element("text-box")
+            .element(ELEMENT_TEXT_BOX)
             .text("")
             .foreground(colors::LINK_WATER_COLOR)
             .font_size(fonts::FONT_SIZE_12)
             .font("Roboto Regular")
             .text_selection(TextSelection::default())
-            .scroll_offset(0.0)
             .padding(4.0)
             .background(colors::LYNCH_COLOR)
             .border_brush("transparent")
@@ -316,7 +389,6 @@ impl Template for TextBox {
             .min_width(128.0)
             .height(32.0)
             .focused(false)
-            .delta(0.0)
             .lost_focus_on_activation(true)
             .child(
                 MouseBehavior::create()
@@ -324,7 +396,7 @@ impl Template for TextBox {
                     .enabled(id)
                     .on_mouse_down(move |states, p| {
                         states
-                            .get::<TextBoxState>(id)
+                            .get_mut::<TextBoxState>(id)
                             .action(TextBoxAction::Mouse(p));
                         true
                     })
@@ -337,39 +409,19 @@ impl Template for TextBox {
                             .padding(id)
                             .child(
                                 Grid::create()
-                                    .child(
-                                        ScrollViewer::create()
-                                            .id("scroll_viewer")
-                                            .scroll_offset(id)
-                                            .scroll_viewer_mode(("custom", "disabled"))
-                                            .delta(id)
-                                            .child(
-                                                TextBlock::create()
-                                                    .id("text_block")
-                                                    .vertical_alignment("center")
-                                                    .foreground(id)
-                                                    .text(id)
-                                                    .water_mark(id)
-                                                    .font(id)
-                                                    .font_size(id)
-                                                    .build(ctx),
-                                            )
-                                            .build(ctx),
-                                    )
+                                    .clip(true)
+                                    // It is important that cursor is the first child
+                                    // should be refactored in the future.
                                     .child(
                                         Cursor::create()
-                                            .element("cursor")
-                                            .id("cursor")
-                                            .margin(0.0)
+                                            .id(ID_CURSOR)
                                             .horizontal_alignment("start")
-                                            .text(id)
-                                            .font(id)
-                                            .font_size(id)
-                                            .scroll_offset(id)
+                                            .text_block(text_block.0)
                                             .focused(id)
                                             .text_selection(id)
                                             .build(ctx),
                                     )
+                                    .child(text_block)
                                     .build(ctx),
                             )
                             .build(ctx),
@@ -378,7 +430,7 @@ impl Template for TextBox {
             )
             .on_key_down(move |states, event| -> bool {
                 states
-                    .get::<TextBoxState>(id)
+                    .get_mut::<TextBoxState>(id)
                     .action(TextBoxAction::Key(event));
                 false
             })